refactor: extract the fact-pack builder in-app — buildSituationPack (Move 2 · P1)
The analyst's fact pack lived only in the CLI scripts/situation-pack.ts, so the in-app
analyst service (P2) couldn't build it. Extract the full builder into
src/lib/growth/situationPack.ts:buildSituationPack(propertyId, asOf, {generator}); the
CLI is now a thin wrapper (parse args, load META_ADS_TOKENS for local dev, call the lib,
write). The situation-analyst skill + backtests keep working unchanged.

Faithful extraction via param-aliasing (PROPERTY/AS_OF kept as local aliases) so the
~600-line body is verbatim; the only intended changes are: params, a parametrized
meta.generator (CLI passes 'scripts/situation-pack.ts'), and the execSync token-fallback
moved to the CLI wrapper (in-app the token is in the runtime env).

Verified: the full pack JSON is BYTE-IDENTICAL before vs after (diff clean), and next
build compiles the new lib cleanly. No app runtime path imports it yet — this unlocks P2
(the in-app analyst service).
